+++ modules/mpd.c
@@ -39,6 +39,7 @@ struct private
     bool repeat;
     bool random;
     bool consume;
+    bool single;
     int volume;
     char *album;
     char *artist;
@@ -176,6 +177,7 @@ content(struct module *mod)
             tag_new_bool(mod, "repeat", m->repeat),
             tag_new_bool(mod, "random", m->random),
             tag_new_bool(mod, "consume", m->consume),
+            tag_new_bool(mod, "single", m->single),
             tag_new_int_range(mod, "volume", m->volume, 0, 100),
             tag_new_string(mod, "album", m->album),
             tag_new_string(mod, "artist", m->artist),
@@ -187,7 +189,7 @@ content(struct module *mod)
             tag_new_int_realtime(
                 mod, "elapsed", elapsed, 0, m->duration, realtime),
         },
-        .count = 13,
+        .count = 14,
     };
 
     mtx_unlock(&mod->lock);
@@ -336,6 +338,7 @@ update_status(struct module *mod)
     m->repeat = mpd_status_get_repeat(status);
     m->random = mpd_status_get_random(status);
     m->consume = mpd_status_get_consume(status);
+    m->single = mpd_status_get_single_state(status) == MPD_SINGLE_ONESHOT;
     m->volume = mpd_status_get_volume(status);
     m->duration = mpd_status_get_total_time(status) * 1000;
     m->elapsed.value = mpd_status_get_elapsed_ms(status);